changing time being int to float

+++ main/models.py
@@ -5,7 +5,7 @@ from django.core.validators import MaxValueValidator, MinValueValidator
 
 class Score(models.Model):
     username=models.CharField(max_length=30)
-    time=models.IntegerField(validators=[MinValueValidator(0),MaxValueValidator(1000)])
+    time=models.FloatField(validators=[MinValueValidator(0),MaxValueValidator(1000)])
     maze_name=models.CharField(max_length=100)
     def __str__(self):
         return f"{self.username}({self.maze_name}):{self.time}"
